hw5/python_docker_code/bondiana/src/requirements.txt
Flask == 2.0.3
Flask-SQLAlchemy == 2.5.1
flask-marshmallow == 0.12.0
marshmallow-sqlalchemy == 0.24.2
psycopg2 == 2.9.2
kubernetes == 20.13.0
PyJWT==2.3.0
